Railsify destroy
