+++ ui/app/controllers/collectors/show.js
@@ -1,15 +1,4 @@
 import Ember from 'ember';
 
 export default Ember.Controller.extend({
-  actions: {
-    destroy: function() {
-      var me = this;
-      var model = this.model;
-
-      this.store.deleteRecord(model);
-      model.save().then(function() {
-        me.transitionToRoute("collectors/index");
-      });
-    }
-  }
 });
