+++ ui/app/routes/collectors/destroy.js
@@ -1,0 +1,7 @@
+import Ember from 'ember';
+
+export default Ember.Route.extend({
+  model: function (params) {
+    return this.store.find('collector', params.id);
+  }
+});
