+++ ui/app/templates/_menu.hbs
@@ -2,7 +2,7 @@
   {{#link-to "dashboard"}}Dashboard{{/link-to}}
 {{/link-to}}
 
-{{#link-to "collectors/index" tagName="li" href=false current-when="collectors/index collectors/new collectors/new_hg collectors/show"}}
+{{#link-to "collectors/index" tagName="li" href=false current-when="collectors/index collectors/new collectors/new_hg collectors/show collectors/destroy"}}
   {{#link-to "collectors/index"}}Collectors{{/link-to}}
 {{/link-to}}
